Fix history list, displays 11 item instead of 10.
Cause:
Checking of how many items are currently on the history, comes first,
before the newly executed command is added to the history.

Fix:
Execute the command first, add the executed command to history, then
check if how many items are already in the history, before
print_history_list() is executed.